+++ tests/test_pdf.py
@@ -111,7 +111,7 @@ class TestPasswords:
 
 class TestPermissions:
     def test_some_permissions_missing(self, resources):
-        with Pdf.open(resources / 'graph-encrypted.pdf', 'owner') as pdf:
+        with Pdf.open(resources / 'graph-encrypted.pdf', password='owner') as pdf:
             assert pdf.allow.print_highres == pdf.allow.modify_annotation == False
             assert pdf.allow.print_lowres